Consolidate 3 PCC test files into single test_pcc.py
Merges test_pcc_validation.py (standalone script, no pytest tests),
test_per_module_pcc.py, and test_per_op_pcc.py into one file.
Deduplicates compute_pcc/assert_with_pcc helpers. All op-level tests
now use the shared device fixture instead of manual open/close.
