--- models/experimental/openvoice/tests/pcc/test_pcc_validation.py
@@ -1,248 +1,0 @@
-#!/usr/bin/env python3
-# SPDX-FileCopyrightText: (c) 2026 Tenstorrent AI
-# SPDX-License-Identifier: Apache-2.0
-
-"""
-PCC (Pearson Correlation Coefficient) validation test.
-
-Compares TTNN outputs against PyTorch reference to validate correctness.
-Target: > 95% correlation (bounty requirement).
-
-IMPORTANT: These tests use deterministic settings (fixed seeds, tau=0.0) to
-ensure reproducible comparisons. In production, the model uses stochastic
-sampling for better audio quality.
-"""
-
-import sys
-
-import numpy as np
-import pytest
-import torch
-
-import ttnn
-
-
-
-def pearson_correlation(a: np.ndarray, b: np.ndarray) -> float:
-    """Calculate Pearson correlation coefficient between two arrays."""
-    a_flat = a.flatten().astype(np.float64)
-    b_flat = b.flatten().astype(np.float64)
-
-    if len(a_flat) != len(b_flat):
-        print(f"Shape mismatch: {a.shape} vs {b.shape}")
-        return 0.0
-
-    if np.std(a_flat) == 0 or np.std(b_flat) == 0:
-        return 1.0 if np.allclose(a_flat, b_flat) else 0.0
-
-    correlation = np.corrcoef(a_flat, b_flat)[0, 1]
-    return float(correlation) if not np.isnan(correlation) else 0.0
-
-
-def validate_voice_conversion():
-    """Validate voice conversion TTNN vs PyTorch."""
-    print("=" * 60)
-    print("Voice Conversion PCC Validation")
-    print("=" * 60)
-
-    import json
-    from pathlib import Path
-
-    checkpoint_dir = Path("checkpoints/converter")
-    if not checkpoint_dir.exists():
-        print("Converter checkpoint not found")
-        return None
-
-    config_path = checkpoint_dir / "config.json"
-    checkpoint_path = checkpoint_dir / "checkpoint.pth"
-
-    with open(config_path) as f:
-        config = json.load(f)
-
-    state_dict = torch.load(checkpoint_path, map_location="cpu")
-    if "model" in state_dict:
-        state_dict = state_dict["model"]
-
-    # Create test input
-    batch_size = 1
-    spec_channels = config["data"]["filter_length"] // 2 + 1
-    seq_len = 100
-
-    torch.manual_seed(42)
-    spec = torch.randn(batch_size, spec_channels, seq_len)
-    spec_lengths = torch.LongTensor([seq_len])
-    sid = torch.zeros(batch_size, 256, 1)
-
-    # Run PyTorch reference (CPU)
-    from models.experimental.openvoice.tt.synthesizer import TTNNSynthesizerTrn
-
-    print("Building CPU model...")
-    model_cpu = TTNNSynthesizerTrn.from_state_dict(state_dict, config, device=None)
-
-    # Use tau=0.0 for deterministic comparison (no stochastic sampling)
-    print("Running PyTorch reference (tau=0.0 for deterministic)...")
-    torch.manual_seed(12345)  # Fixed seed for any residual randomness
-    with torch.no_grad():
-        audio_cpu, _, _ = model_cpu.voice_conversion(spec, spec_lengths, sid, sid, tau=0.0)
-    if audio_cpu.dtype == torch.bfloat16:
-        audio_cpu = audio_cpu.float()
-    audio_cpu_np = audio_cpu.numpy()
-    print(f"  Output shape: {audio_cpu_np.shape}")
-
-    # Run TTNN
-    print("Running TTNN...")
-    device = ttnn.open_device(device_id=0)
-    try:
-        print("Building TTNN model...")
-        model_ttnn = TTNNSynthesizerTrn.from_state_dict(state_dict, config, device=device)
-
-        spec_ttnn = ttnn.from_torch(spec.float(), dtype=ttnn.bfloat16, device=device)
-        spec_lengths_ttnn = ttnn.from_torch(spec_lengths, dtype=ttnn.int32, device=device)
-        sid_ttnn = ttnn.from_torch(sid.float(), dtype=ttnn.bfloat16, device=device)
-
-        torch.manual_seed(12345)  # Same seed
-        with torch.no_grad():
-            audio_ttnn, _, _ = model_ttnn.voice_conversion(spec_ttnn, spec_lengths_ttnn, sid_ttnn, sid_ttnn, tau=0.0)
-
-        if not isinstance(audio_ttnn, torch.Tensor):
-            audio_ttnn = ttnn.to_torch(ttnn.from_device(audio_ttnn))
-        if audio_ttnn.dtype == torch.bfloat16:
-            audio_ttnn = audio_ttnn.float()
-        audio_ttnn_np = audio_ttnn.numpy()
-        print(f"  Output shape: {audio_ttnn_np.shape}")
-
-        pcc = pearson_correlation(audio_cpu_np, audio_ttnn_np)
-        print(f"\nPCC (Voice Conversion): {pcc:.4f}")
-        print(f"Target: > 0.95")
-        print(f"Status: {'PASS' if pcc > 0.95 else 'FAIL'}")
-
-        return pcc
-    finally:
-        ttnn.close_device(device)
-
-
-def validate_tts():
-    """Validate TTS TTNN vs PyTorch."""
-    print("\n" + "=" * 60)
-    print("TTS PCC Validation")
-    print("=" * 60)
-
-    import json
-    from pathlib import Path
-
-    checkpoint_dir = Path("checkpoints/melo/EN")
-    if not checkpoint_dir.exists():
-        print("MeloTTS checkpoint not found")
-        return None
-
-    config_path = checkpoint_dir / "config.json"
-    checkpoint_path = checkpoint_dir / "checkpoint.pth"
-
-    with open(config_path) as f:
-        config = json.load(f)
-
-    state_dict = torch.load(checkpoint_path, map_location="cpu")
-    if "model" in state_dict:
-        state_dict = state_dict["model"]
-
-    # Create test input
-    n_symbols = len(config.get("symbols", [])) or 200
-    token_len = 15
-
-    torch.manual_seed(42)
-    tokens = torch.randint(0, n_symbols, (1, token_len))
-    tones = torch.zeros(1, token_len, dtype=torch.long)
-    lang_ids = torch.zeros(1, token_len, dtype=torch.long)
-    bert = torch.randn(1, 1024, token_len)
-    ja_bert = torch.randn(1, 768, token_len)
-    sid = torch.LongTensor([0])
-    x_lengths = torch.LongTensor([token_len])
-
-    # Build model (shared between CPU and TTNN since _infer_ttnn falls back to PyTorch)
-    from models.experimental.openvoice.tt.melo_tts import TTNNMeloTTS
-
-    model = TTNNMeloTTS.from_state_dict(state_dict, config, device=None)
-
-    # Run PyTorch reference
-    print("Running PyTorch reference...")
-    torch.manual_seed(12345)  # Fixed seed for deterministic sampling
-    with torch.no_grad():
-        audio_cpu, _, _ = model.infer(tokens, x_lengths, sid, tones, lang_ids, bert, ja_bert)
-    if audio_cpu.dtype == torch.bfloat16:
-        audio_cpu = audio_cpu.float()
-    audio_cpu_np = audio_cpu.numpy()
-    print(f"  Output shape: {audio_cpu_np.shape}")
-
-    # Run TTNN (which internally uses _infer_ttnn -> _infer_pytorch)
-    print("Running TTNN...")
-    device = ttnn.open_device(device_id=0)
-    try:
-        model_ttnn = TTNNMeloTTS.from_state_dict(state_dict, config, device=device)
-
-        tokens_ttnn = ttnn.from_torch(tokens, dtype=ttnn.int32, device=device)
-        tones_ttnn = ttnn.from_torch(tones, dtype=ttnn.int32, device=device)
-        lang_ids_ttnn = ttnn.from_torch(lang_ids, dtype=ttnn.int32, device=device)
-        bert_ttnn = ttnn.from_torch(bert.float(), dtype=ttnn.bfloat16, device=device)
-        ja_bert_ttnn = ttnn.from_torch(ja_bert.float(), dtype=ttnn.bfloat16, device=device)
-        sid_ttnn = ttnn.from_torch(sid, dtype=ttnn.int32, device=device)
-        x_lengths_ttnn = ttnn.from_torch(x_lengths, dtype=ttnn.int32, device=device)
-
-        # Same seed for deterministic comparison
-        torch.manual_seed(12345)
-        with torch.no_grad():
-            audio_ttnn, _, _ = model_ttnn.infer(
-                tokens_ttnn, x_lengths_ttnn, sid_ttnn, tones_ttnn, lang_ids_ttnn, bert_ttnn, ja_bert_ttnn
-            )
-
-        if not isinstance(audio_ttnn, torch.Tensor):
-            audio_ttnn = ttnn.to_torch(ttnn.from_device(audio_ttnn))
-        if audio_ttnn.dtype == torch.bfloat16:
-            audio_ttnn = audio_ttnn.float()
-        audio_ttnn_np = audio_ttnn.numpy()
-        print(f"  Output shape: {audio_ttnn_np.shape}")
-
-        pcc = pearson_correlation(audio_cpu_np, audio_ttnn_np)
-        print(f"\nPCC (TTS): {pcc:.4f}")
-        print(f"Target: > 0.95")
-        print(f"Status: {'PASS' if pcc > 0.95 else 'FAIL'}")
-
-        return pcc
-    finally:
-        ttnn.close_device(device)
-
-
-def main():
-    print("OpenVoice TTNN - PCC Validation Test")
-    print("Target: Token-level accuracy > 95%")
-    print("Note: Using deterministic settings for reproducible comparison")
-    print()
-
-    results = {}
-
-    # Validate voice conversion
-    vc_pcc = validate_voice_conversion()
-    if vc_pcc is not None:
-        results["voice_conversion"] = vc_pcc
-
-    # Validate TTS
-    tts_pcc = validate_tts()
-    if tts_pcc is not None:
-        results["tts"] = tts_pcc
-
-    # Summary
-    print("\n" + "=" * 60)
-    print("SUMMARY")
-    print("=" * 60)
-    all_pass = True
-    for name, pcc in results.items():
-        status = "PASS" if pcc > 0.95 else "FAIL"
-        if pcc <= 0.95:
-            all_pass = False
-        print(f"  {name}: PCC = {pcc:.4f} [{status}]")
-
-    print(f"\nOverall: {'ALL TESTS PASSED' if all_pass else 'SOME TESTS FAILED'}")
-    return 0 if all_pass else 1
-
-
-if __name__ == "__main__":
-    sys.exit(main())
